Moved log files to separate module called log. Added methods so that the logger can be explicitly called with debug or error message.
