--- src/test/loggingTest.conf
@@ -1,25 +1,0 @@
-[loggers]
-keys=root
-
-[handlers]
-keys=timedRotatingFileHandler
-
-[formatters]
-keys=simpleFormatter
-
-[logger_root]
-level=DEBUG
-handlers=timedRotatingFileHandler
-
-[handler_timedRotatingFileHandler]
-class=logging.handlers.TimedRotatingFileHandler
-level=DEBUG
-d='loggingTest-' + __date__ + '.log'
-args=('%(logfilename)s','S', 10)
-formatter=simpleFormatter
-
-[formatter_simpleFormatter]
-format=%(asctime)s - %(name)s - %(levelname)s - %(message)s
-datefmt=%Y-%m-%d:%H:%M:%S
-
-
